+++ bittensor_cli/cli.py
@@ -8,6 +8,7 @@ import re
 import ssl
 import sys
 import traceback
+import warnings
 from pathlib import Path
 from typing import Coroutine, Optional
 from dataclasses import fields
@@ -71,8 +72,6 @@ from bittensor_cli.src.bittensor.utils import (
     prompt_for_identity,
     validate_uri,
     prompt_for_subnet_identity,
-    print_linux_dependency_message,
-    is_linux,
     validate_rate_tolerance,
 )
 
@@ -264,14 +263,15 @@ class Options:
         "--slippage-tolerance",
         "--tolerance",
         "--rate-tolerance",
-        help="Set the rate tolerance percentage for transactions (default: 0.05%).",
+        help="Set the rate tolerance percentage for transactions (default: 0.05 for 5%).",
         callback=validate_rate_tolerance,
     )
     safe_staking = typer.Option(
         None,
         "--safe-staking/--no-safe-staking",
         "--safe/--unsafe",
-        help="Enable or disable safe staking mode (default: enabled).",
+        show_default=False,
+        help="Enable or disable safe staking mode [dim](default: enabled)[/dim].",
     )
     allow_partial_stake = typer.Option(
         None,
@@ -279,7 +279,8 @@ class Options:
         "--partial/--no-partial",
         "--allow/--not-allow",
         "--allow-partial/--not-partial",
-        help="Enable or disable partial stake mode (default: disabled).",
+        show_default=False,
+        help="Enable or disable partial stake mode [dim](default: disabled)[/dim].",
     )
     dashboard_path = typer.Option(
         None,
@@ -296,8 +297,11 @@ class Options:
         "--json-out",
         help="Outputs the result of the command as JSON.",
     )
-    era: int = typer.Option(
-        3, help="Length (in blocks) for which the transaction should be valid."
+    period: int = typer.Option(
+        16,
+        "--period",
+        "--era",
+        help="Length (in blocks) for which the transaction should be valid.",
     )
 
 
@@ -439,36 +443,49 @@ def parse_mnemonic(mnemonic: str) -> str:
 def get_creation_data(
     mnemonic: Optional[str],
     seed: Optional[str],
-    json: Optional[str],
+    json_path: Optional[str],
     json_password: Optional[str],
 ) -> tuple[str, str, str, str]:
     """
     Determines which of the key creation elements have been supplied, if any. If None have been supplied,
     prompts to user, and determines what they've supplied. Returns all elements in a tuple.
     """
-    if not mnemonic and not seed and not json:
-        prompt_answer = Prompt.ask(
-            "Enter the mnemonic, or the seed hex string, or the location of the JSON file."
+    if not mnemonic and not seed and not json_path:
+        choices = {
+            1: "mnemonic",
+            2: "seed hex string",
+            3: "path to JSON File",
+        }
+        type_answer = IntPrompt.ask(
+            "Select one of the following to enter\n"
+            f"[{COLORS.G.HINT}][1][/{COLORS.G.HINT}] Mnemonic\n"
+            f"[{COLORS.G.HINT}][2][/{COLORS.G.HINT}] Seed hex string\n"
+            f"[{COLORS.G.HINT}][3][/{COLORS.G.HINT}] Path to JSON File\n",
+            choices=["1", "2", "3"],
+            show_choices=False,
         )
-        if prompt_answer.startswith("0x"):
+        prompt_answer = Prompt.ask(f"Please enter your {choices[type_answer]}")
+        if type_answer == 1:
+            mnemonic = prompt_answer
+        elif type_answer == 2:
             seed = prompt_answer
-        elif len(prompt_answer.split(" ")) > 1:
-            mnemonic = parse_mnemonic(prompt_answer)
-        else:
-            json = prompt_answer
+            if seed.startswith("0x"):
+                seed = seed[2:]
+        elif type_answer == 3:
+            json_path = prompt_answer
     elif mnemonic:
         mnemonic = parse_mnemonic(mnemonic)
 
-    if json:
-        if not os.path.exists(json):
-            print_error(f"The JSON file '{json}' does not exist.")
+    if json_path:
+        if not os.path.exists(json_path):
+            print_error(f"The JSON file '{json_path}' does not exist.")
             raise typer.Exit()
 
-    if json and not json_password:
+    if json_path and not json_password:
         json_password = Prompt.ask(
             "Enter the backup password for JSON file.", password=True
         )
-    return mnemonic, seed, json, json_password
+    return mnemonic, seed, json_path, json_password
 
 
 def config_selector(conf: dict, title: str):
@@ -899,6 +916,12 @@ class CLIManager:
         self.subnets_app.command(
             "get-identity", rich_help_panel=HELP_PANELS["SUBNETS"]["IDENTITY"]
         )(self.subnets_get_identity)
+        self.subnets_app.command(
+            "start", rich_help_panel=HELP_PANELS["SUBNETS"]["CREATION"]
+        )(self.subnets_start)
+        self.subnets_app.command(
+            "check-start", rich_help_panel=HELP_PANELS["SUBNETS"]["INFO"]
+        )(self.subnets_check_start)
 
         # weights commands
         self.weights_app.command(
@@ -1002,31 +1025,37 @@ class CLIManager:
         :param network: Network name (e.g. finney, test, etc.) or
                         chain endpoint (e.g. ws://127.0.0.1:9945, wss://entrypoint-finney.opentensor.ai:443)
         """
-        if not self.subtensor:
-            if network:
-                network_ = None
-                for item in network:
-                    if item.startswith("ws"):
-                        network_ = item
-                        break
-                    else:
-                        network_ = item
+        with warnings.catch_warnings():
+            warnings.filterwarnings(
+                "ignore",
+                "You are instantiating the AsyncSubstrateInterface Websocket outside of an event loop. "
+                "Verify this is intended.",
+            )
+            if not self.subtensor:
+                if network:
+                    network_ = None
+                    for item in network:
+                        if item.startswith("ws"):
+                            network_ = item
+                            break
+                        else:
+                            network_ = item
 
-                not_selected_networks = [net for net in network if net != network_]
-                if not_selected_networks:
+                    not_selected_networks = [net for net in network if net != network_]
+                    if not_selected_networks:
+                        console.print(
+                            f"Networks not selected: [dark_orange]{', '.join(not_selected_networks)}[/dark_orange]"
+                        )
+
+                    self.subtensor = SubtensorInterface(network_)
+                elif self.config["network"]:
+                    self.subtensor = SubtensorInterface(self.config["network"])
                     console.print(
-                        f"Networks not selected: [dark_orange]{', '.join(not_selected_networks)}[/dark_orange]"
+                        f"Using the specified network [{COLORS.G.LINKS}]{self.config['network']}"
+                        f"[/{COLORS.G.LINKS}] from config"
                     )
-
-                self.subtensor = SubtensorInterface(network_)
-            elif self.config["network"]:
-                self.subtensor = SubtensorInterface(self.config["network"])
-                console.print(
-                    f"Using the specified network [{COLORS.G.LINKS}]{self.config['network']}"
-                    f"[/{COLORS.G.LINKS}] from config"
-                )
-            else:
-                self.subtensor = SubtensorInterface(defaults.subtensor.network)
+                else:
+                    self.subtensor = SubtensorInterface(defaults.subtensor.network)
         return self.subtensor
 
     def _run_command(self, cmd: Coroutine, exit_early: bool = True):
@@ -1210,12 +1239,14 @@ class CLIManager:
             "--safe-staking/--no-safe-staking",
             "--safe/--unsafe",
             help="Enable or disable safe staking mode.",
+            show_default=False,
         ),
         allow_partial_stake: Optional[bool] = typer.Option(
             None,
             "--allow-partial-stake/--no-allow-partial-stake",
             "--partial/--no-partial",
             "--allow/--not-allow",
+            show_default=False,
         ),
         dashboard_path: Optional[str] = Options.dashboard_path,
     ):
@@ -1808,7 +1839,7 @@ class CLIManager:
         transfer_all: bool = typer.Option(
             False, "--all", prompt=False, help="Transfer all available balance."
         ),
-        era: int = Options.era,
+        period: int = Options.period,
         wallet_name: str = Options.wallet_name,
         wallet_path: str = Options.wallet_path,
         wallet_hotkey: str = Options.wallet_hotkey,
@@ -1863,7 +1894,7 @@ class CLIManager:
                 destination=destination_ss58_address,
                 amount=amount,
                 transfer_all=transfer_all,
-                era=era,
+                era=period,
                 prompt=prompt,
                 json_output=json_output,
             )
@@ -2106,7 +2137,7 @@ class CLIManager:
         wallet_hotkey: Optional[str] = Options.wallet_hotkey,
         mnemonic: Optional[str] = Options.mnemonic,
         seed: Optional[str] = Options.seed,
-        json: Optional[str] = Options.json,
+        json_path: Optional[str] = Options.json,
         json_password: Optional[str] = Options.json_password,
         use_password: Optional[bool] = Options.use_password,
         overwrite: bool = Options.overwrite,
@@ -2146,15 +2177,15 @@ class CLIManager:
 
         wallet = Wallet(wallet_name, wallet_hotkey, wallet_path)
 
-        mnemonic, seed, json, json_password = get_creation_data(
-            mnemonic, seed, json, json_password
+        mnemonic, seed, json_path, json_password = get_creation_data(
+            mnemonic, seed, json_path, json_password
         )
         return self._run_command(
             wallets.regen_coldkey(
                 wallet,
                 mnemonic,
                 seed,
-                json,
+                json_path,
                 json_password,
                 use_password,
                 overwrite,
@@ -2230,7 +2261,7 @@ class CLIManager:
         wallet_hotkey: Optional[str] = Options.wallet_hotkey,
         mnemonic: Optional[str] = Options.mnemonic,
         seed: Optional[str] = Options.seed,
-        json: Optional[str] = Options.json,
+        json_path: Optional[str] = Options.json,
         json_password: Optional[str] = Options.json_password,
         use_password: bool = typer.Option(
             False,  # Overriden to False
@@ -2266,15 +2297,15 @@ class CLIManager:
             ask_for=[WO.NAME, WO.PATH, WO.HOTKEY],
             validate=WV.WALLET,
         )
-        mnemonic, seed, json, json_password = get_creation_data(
-            mnemonic, seed, json, json_password
+        mnemonic, seed, json_path, json_password = get_creation_data(
+            mnemonic, seed, json_path, json_password
         )
         return self._run_command(
             wallets.regen_hotkey(
                 wallet,
                 mnemonic,
                 seed,
-                json,
+                json_path,
                 json_password,
                 use_password,
                 overwrite,
@@ -3206,7 +3237,7 @@ class CLIManager:
         rate_tolerance: Optional[float] = Options.rate_tolerance,
         safe_staking: Optional[bool] = Options.safe_staking,
         allow_partial_stake: Optional[bool] = Options.allow_partial_stake,
-        era: int = Options.era,
+        period: int = Options.period,
         prompt: bool = Options.prompt,
         quiet: bool = Options.quiet,
         verbose: bool = Options.verbose,
@@ -3402,7 +3433,7 @@ class CLIManager:
                 rate_tolerance,
                 allow_partial_stake,
                 json_output,
-                era,
+                period,
             )
         )
 
@@ -3454,7 +3485,7 @@ class CLIManager:
         rate_tolerance: Optional[float] = Options.rate_tolerance,
         safe_staking: Optional[bool] = Options.safe_staking,
         allow_partial_stake: Optional[bool] = Options.allow_partial_stake,
-        era: int = Options.era,
+        period: int = Options.period,
         prompt: bool = Options.prompt,
         interactive: bool = typer.Option(
             False,
@@ -3647,7 +3678,7 @@ class CLIManager:
                     exclude_hotkeys=exclude_hotkeys,
                     prompt=prompt,
                     json_output=json_output,
-                    era=era,
+                    era=period,
                 )
             )
         elif (
@@ -3703,7 +3734,7 @@ class CLIManager:
                 rate_tolerance=rate_tolerance,
                 allow_partial_stake=allow_partial_stake,
                 json_output=json_output,
-                era=era,
+                era=period,
             )
         )
 
@@ -3731,7 +3762,7 @@ class CLIManager:
         stake_all: bool = typer.Option(
             False, "--stake-all", "--all", help="Stake all", prompt=False
         ),
-        era: int = Options.era,
+        period: int = Options.period,
         prompt: bool = Options.prompt,
         quiet: bool = Options.quiet,
         verbose: bool = Options.verbose,
@@ -3861,7 +3892,7 @@ class CLIManager:
                 destination_hotkey=destination_hotkey,
                 amount=amount,
                 stake_all=stake_all,
-                era=era,
+                era=period,
                 interactive_selection=interactive_selection,
                 prompt=prompt,
             )
@@ -3902,7 +3933,7 @@ class CLIManager:
         stake_all: bool = typer.Option(
             False, "--stake-all", "--all", help="Stake all", prompt=False
         ),
-        era: int = Options.era,
+        period: int = Options.period,
         prompt: bool = Options.prompt,
         quiet: bool = Options.quiet,
         verbose: bool = Options.verbose,
@@ -4024,7 +4055,7 @@ class CLIManager:
                 dest_netuid=dest_netuid,
                 dest_coldkey_ss58=dest_ss58,
                 amount=amount,
-                era=era,
+                era=period,
                 interactive_selection=interactive_selection,
                 stake_all=stake_all,
                 prompt=prompt,
@@ -4066,7 +4097,7 @@ class CLIManager:
             "--all",
             help="Swap all available stake",
         ),
-        era: int = Options.era,
+        period: int = Options.period,
         prompt: bool = Options.prompt,
         wait_for_inclusion: bool = Options.wait_for_inclusion,
         wait_for_finalization: bool = Options.wait_for_finalization,
@@ -4131,7 +4162,7 @@ class CLIManager:
                 destination_netuid=dest_netuid,
                 amount=amount,
                 swap_all=swap_all,
-                era=era,
+                era=period,
                 interactive_selection=interactive_selection,
                 prompt=prompt,
                 wait_for_inclusion=wait_for_inclusion,
@@ -4591,6 +4622,7 @@ class CLIManager:
         param_value: Optional[str] = typer.Option(
             "", "--value", help="Value to set the hyperparameter to."
         ),
+        prompt: bool = Options.prompt,
         quiet: bool = Options.quiet,
         verbose: bool = Options.verbose,
         json_output: bool = Options.json_output,
@@ -4615,6 +4647,11 @@ class CLIManager:
                 raise typer.Exit()
 
         if not param_name:
+            if not prompt:
+                err_console.print(
+                    "Param name not supplied with `--no-prompt` flag. Cannot continue"
+                )
+                return False
             hyperparam_list = [field.name for field in fields(SubnetHyperparameters)]
             console.print("Available hyperparameters:\n")
             for idx, param in enumerate(hyperparam_list, start=1):
@@ -4628,6 +4665,11 @@ class CLIManager:
             param_name = hyperparam_list[choice - 1]
 
         if param_name in ["alpha_high", "alpha_low"]:
+            if not prompt:
+                err_console.print(
+                    "`alpha_high` and `alpha_low` values cannot be set with `--no-prompt`"
+                )
+                return False
             param_name = "alpha_values"
             low_val = FloatPrompt.ask(
                 "Enter the new value for [dark_orange]alpha_low[/dark_orange]"
@@ -4638,6 +4680,11 @@ class CLIManager:
             param_value = f"{low_val},{high_val}"
 
         if not param_value:
+            if not prompt:
+                err_console.print(
+                    "Param value not supplied with `--no-prompt` flag. Cannot continue."
+                )
+                return False
             if HYPERPARAMS.get(param_name):
                 param_value = Prompt.ask(
                     f"Enter the new value for [{COLORS.G.SUBHEAD}]{param_name}[/{COLORS.G.SUBHEAD}] "
@@ -4656,6 +4703,7 @@ class CLIManager:
                 netuid,
                 param_name,
                 param_value,
+                prompt,
                 json_output,
             )
         )
@@ -4996,9 +5044,6 @@ class CLIManager:
         if all_netuids and not json_output:
             html_output = True
 
-        if html_output and is_linux():
-            print_linux_dependency_message()
-
         return self._run_command(
             price.price(
                 self.initialize_chain(network),
@@ -5143,6 +5188,70 @@ class CLIManager:
         self._run_command(
             subnets.create(
                 wallet, self.initialize_chain(network), identity, json_output, prompt
+            )
+        )
+
+    def subnets_check_start(
+        self,
+        network: Optional[list[str]] = Options.network,
+        netuid: int = Options.netuid,
+        quiet: bool = Options.quiet,
+        verbose: bool = Options.verbose,
+    ):
+        """
+        Checks if a subnet's emission schedule can be started.
+
+        This command verifies if a subnet's emission schedule can be started based on the subnet's registration block.
+
+        Example:
+        [green]$[/green] btcli subnets check_start --netuid 1
+        """
+        self.verbosity_handler(quiet, verbose)
+        return self._run_command(
+            subnets.get_start_schedule(self.initialize_chain(network), netuid)
+        )
+
+    def subnets_start(
+        self,
+        wallet_name: str = Options.wallet_name,
+        wallet_path: str = Options.wallet_path,
+        wallet_hotkey: str = Options.wallet_hotkey,
+        network: Optional[list[str]] = Options.network,
+        netuid: int = Options.netuid,
+        prompt: bool = Options.prompt,
+        quiet: bool = Options.quiet,
+        verbose: bool = Options.verbose,
+    ):
+        """
+        Starts a subnet's emission schedule.
+
+        The owner of the subnet must call this command to start the emission schedule.
+
+        Example:
+        [green]$[/green] btcli subnets start --netuid 1
+        [green]$[/green] btcli subnets start --netuid 1 --wallet-name alice
+        """
+        self.verbosity_handler(quiet, verbose)
+        if not wallet_name:
+            wallet_name = Prompt.ask(
+                "Enter the [blue]wallet name[/blue] [dim](which you used to create the subnet)[/dim]",
+                default=self.config.get("wallet_name") or defaults.wallet.name,
+            )
+        wallet = self.wallet_ask(
+            wallet_name,
+            wallet_path,
+            wallet_hotkey,
+            ask_for=[
+                WO.NAME,
+            ],
+            validate=WV.WALLET,
+        )
+        return self._run_command(
+            subnets.start_subnet(
+                wallet,
+                self.initialize_chain(network),
+                netuid,
+                prompt,
             )
         )
 
@@ -5355,10 +5464,12 @@ class CLIManager:
         wallet_hotkey: str = Options.wallet_hotkey,
         network: Optional[list[str]] = Options.network,
         netuid: int = Options.netuid,
-        era: Optional[
+        period: Optional[
             int
-        ] = typer.Option(  # Should not be Options.era bc this needs to be an Optional[int]
+        ] = typer.Option(  # Should not be Options.period bc this needs to be an Optional[int]
             None,
+            "--period",
+            "--era",
             help="Length (in blocks) for which the transaction should be valid. Note that it is possible that if you "
             "use an era for this transaction that you may pay a different fee to register than the one stated.",
         ),
@@ -5391,7 +5502,7 @@ class CLIManager:
                 wallet,
                 self.initialize_chain(network),
                 netuid,
-                era,
+                period,
                 json_output,
                 prompt,
             )
@@ -5710,7 +5821,7 @@ class CLIManager:
             help="Coldkey SS58 address to view dashboard for",
         ),
         use_wry: bool = typer.Option(
-            False, "--use-wry", help="Use PyWry instead of browser window"
+            False, "--use-wry", "--html", help="Display output in browser window."
         ),
         save_file: bool = typer.Option(
             False, "--save-file", "--save", help="Save the dashboard HTML file"
@@ -5723,12 +5834,10 @@ class CLIManager:
         Display html dashboard with subnets list, stake, and neuron information.
         """
         self.verbosity_handler(quiet, verbose)
-        if use_wry and is_linux():
-            print_linux_dependency_message()
 
         if use_wry and save_file:
-            print_error("Cannot save file when using PyWry.")
-            raise typer.Exit()
+            print_error("Cannot save file when using browser output.")
+            return
 
         if save_file:
             if not dashboard_path: